chore(dev.yml): update dev workflow to use make-jvm-workflow.yml instead of dev-workflow.yml chore(dev.yml): remove commented out code for branch filtering in the dev workflow The dev workflow has been updated to use the make-jvm-workflow.yml instead of the previous dev-workflow.yml. The commented out code for branch filtering has been removed as it is no longer needed.
chore(dev.yml): add docs workflow for publishing Storybook documentation
The dev.yml file now includes a new docs workflow for publishing the Storybook documentation. This workflow uses the publish-storybook-workflow.yml file and has the necessary permissions to read contents, write packages, write pages, and write id-token. The workflow is configured to disable Chromatic integration and specifies the storybook directory and storybook-static directory. The necessary secrets NPM_AUTH_TOKEN and CHROMATIC_PROJECT_TOKEN are also provided.

chore(Dockerfile): update base image to node:21.7-alpine3.19
The base image in the Dockerfile for the storybook has been updated to node:21.7-alpine3.19. This ensures that the latest version of Node.js is used for building the React application.

chore(Dockerfile): update nginx base image to nginx:1.25.4-alpine
The base image for the nginx container in the Dockerfile has been updated to nginx:1.25.4-alpine. This ensures that the latest version of nginx is used.

chore(Dockerfile): update nginx container permissions
The permissions for the nginx container in the Dockerfile have been updated to ensure that the nginx user owns the necessary directories and files.

chore(manager.js): update logo import path
The import path for the logo image in the manager.js file has been updated to reflect the new location of the logo image.

chore(preview.js): update ThemeContextProvider and StorybookCanvas import paths
The import paths for the ThemeContextProvider and StorybookCanvas components in the preview.js file have been updated to reflect the new package names.

chore(webpack.config.js): remove unused code
The unused code for customizing the renderer in the webpack.config.js file has been removed.

chore(package.json): update dependencies and scripts for Storybook
The dependencies and scripts in the package.json file for the Storybook have been updated to use the latest versions. The @komune-io/g2 and @komune-io/g2-storybook-document