feat(infra): update docker-compose configuration for BCLAN services chore(infra): improve readability and maintainability of docker-compose files
The changes in the .env_dev file include updating the DOCKER_NETWORK variable to 's2-test-network' and adding new environment variables for BCLAN CouchDB configuration. This update allows for better organization and clarity in the configuration of the BCLAN services.

In the docker-compose files, the version has been updated to '3.8' for compatibility. Service names and configurations have been adjusted to use the updated DOCKER_NETWORK variable and VERSION_C2. Ports have been specified for certain services, and dependencies have been added for peer0.bc-coop.bclan service. These changes enhance the consistency and coherence of the docker-compose setup for BCLAN services.