feat(S2Automate.kt): add support for withResultAsAction property in S2Automate class to track if transitions have results refactor(SsmExtention.kt): update toSsmTransitions and toSsmTransition functions to use withResultAsAction property refactor(SsmAutomatePersister.kt): introduce withEventAsAction variable to handle event as action based on withResultAsAction property feat(OrderBookDeciderEventImpl.kt): implement OrderBookDecider interface with order book creation, update, publish, and close decisions feat(OrderBookAutomateConfig.kt): rename automate function to automate2 and update the name property in the s2 block to include a unique identifier
The changes were made to enhance the functionality of the S2Automate, SsmExtention, SsmAutomatePersister, OrderBookDeciderEventImpl, and OrderBookAutomateConfig classes by introducing a new property to track if transitions have results, updating functions to utilize this property, and improving naming conventions for better clarity and organization.

feat(orderbook-storing): add OrderBookAutomateEventConfig and OrderBookStoringDeciderEventImplTest classes
The OrderBookAutomateEventConfig class is added to configure the automation of order book events in the storing application. This class sets up the automation process for creating, updating, publishing, and closing order books. It also specifies the chaincode URI and signer agent for the automation.

The OrderBookStoringDeciderEventImplTest class is added to test the functionality of the order book storing decider events. It includes tests for creating order books, updating them, publishing them, and closing them. Additionally, it tests the flow of events to build entities and verifies the status of the order books after each operation.